Extend Scala 2.13 support to most interpreters. Drop 2.11 support in most interpreters as well.
Update akka-http to 10.1.9
Update akka-actor to 2.5.25
Update akka-http-json to 1.28.0
Update play to 2.7.3. Remove endpoints.play.server.HttpServer and simplify endpoints.play.server.PlayComponents.
Update sttp to 1.6.6
Update circe to 0.12.1. A consequence is that most modules have to drop Scala 2.11 support (because it is not anymore supported by circe)
Update Scala.js to 0.6.29
Update sbt to 1.3.2