Add streaming support
Fixes #6

Introduces a `ChunkedEntities` algebra defining request and response entities that use the chunked transfer-encoding.

Streamed entities are represented by a type `Chunks[A]` in the algebra. For instance, a streamed response entity containing chunks of type `Event` has type `ResponseEntity[Chunks[Event]]`, and an endpoint returning such an entity has type `Endpoint[Unit, Chunks[Event]]`.

The type `Chunks[A]` is an abstract type, which is fixed by interpreters. Both client and server interpreters currently fix it to Akka Stream’s `Source[A, _]`.

Since Wiremock server does not support streaming, we use Akka-HTTP in tests to implement either a server (when testing client interpreters) or a client (when testing a server interpreter).

The type `ResponseEntity` of the Play server interpreter had to be adapted to support chunked transfer-encoding.